# proyecto/lightning-mcfilm/api-gateway/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Lightning McFilm API
  description: |
    API Gateway unificado para Lightning McFilm

    **Microservicios:**
    - **Auth Service** → http://localhost:8000
    - **Movie Service** → http://localhost:4000/api/movies
    - **Gateway** → http://localhost:4000

    ¡Ka-chow!
  version: 2.2.1
servers:
  - url: http://localhost:4000
    description: Gateway Local
  - url: http://localhost:8000
    description: Auth Service Directo (para login/logout)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/movies/search:
    get:
      tags: [Movie Service]
      summary: Buscar películas
      parameters:
        - name: s
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: Resultados de búsqueda }
  /api/movies/detail/{imdbID}:
    get:
      tags: [Movie Service]
      summary: Detalle completo de película
      parameters:
        - name: imdbID
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Película completa con todos los datos }
...